+++ libs/common/src/rbac/index.ts
@@ -1,0 +1,3 @@
+export * from './system-rollen.const';
+export * from './berechtigung.decorator';
+export * from './rbac.guard';